refactor: разбивает main.js по-модулям, слоям
