+++ js/entity/user/names.js
@@ -1,0 +1,6 @@
+import { entityDocumentFromArray } from '../../shared/db/entity-document.js';
+
+const APP_NAMES_BANK = [{ firstName: 'Terry', lastName: 'Medhurst', maidenName: 'Smitham' }, { firstName: 'Sheldon', lastName: 'Quigley', maidenName: 'Cole' }, { firstName: 'Terrill', lastName: 'Hills', maidenName: 'Hoeger' }, { firstName: 'Miles', lastName: 'Cummerata', maidenName: 'Maggio' }, { firstName: 'Mavis', lastName: 'Schultz', maidenName: 'Yundt' }, { firstName: 'Alison', lastName: 'Reichert', maidenName: 'Franecki' }, { firstName: 'Oleta', lastName: 'Abbott', maidenName: 'Wyman' }, { firstName: 'Ewell', lastName: 'Mueller', maidenName: 'Durgan' }, { firstName: 'Demetrius', lastName: 'Corkery', maidenName: 'Gleason' }, { firstName: 'Eleanora', lastName: 'Price', maidenName: 'Cummings' }, { firstName: 'Marcel', lastName: 'Jones', maidenName: 'Smith' }, { firstName: 'Assunta', lastName: 'Rath', maidenName: 'Heller' }, { firstName: 'Trace', lastName: 'Douglas', maidenName: 'Lemke' }, { firstName: 'Enoch', lastName: 'Lynch', maidenName: 'Heidenreich' }, { firstName: 'Jeanne', lastName: 'Halvorson', maidenName: 'Cummerata' }, { firstName: 'Trycia', lastName: 'Fadel', maidenName: 'Rosenbaum' }, { firstName: 'Bradford', lastName: 'Prohaska', maidenName: 'Bins' }, { firstName: 'Arely', lastName: 'Skiles', maidenName: 'Monahan' }, { firstName: 'Gust', lastName: 'Purdy', maidenName: 'Abshire' }, { firstName: 'Lenna', lastName: 'Renner', maidenName: 'Schumm' }, { firstName: 'Doyle', lastName: 'Ernser', maidenName: 'Feeney' }, { firstName: 'Tressa', lastName: 'Weber', maidenName: 'Williamson' }, { firstName: 'Felicity', lastName: 'O`Reilly', maidenName: 'Rosenbaum' }, { firstName: 'Jocelyn', lastName: 'Schuster', maidenName: 'Dooley' }, { firstName: 'Edwina', lastName: 'Ernser', maidenName: 'Kiehn' }, { firstName: 'Griffin', lastName: 'Braun', maidenName: 'Deckow' }, { firstName: 'Piper', lastName: 'Schowalter', maidenName: 'Wuckert' }, { firstName: 'Kody', lastName: 'Terry', maidenName: 'Larkin' }, { firstName: 'Macy', lastName: 'Greenfelder', maidenName: 'Koepp' }, { firstName: 'Maurine', lastName: 'Stracke', maidenName: 'Abshire' }];
+
+
+export const createNamesDocument = () => entityDocumentFromArray(APP_NAMES_BANK);
