+++ js/keks-db/config.js
@@ -1,0 +1,9 @@
+
+const APP_PHOTO_COUNT = 25;
+const APP_USERS_COUNT = 6;
+const APP_COMMENTS_FOR_POST = 5;
+export {
+  APP_PHOTO_COUNT,
+  APP_USERS_COUNT,
+  APP_COMMENTS_FOR_POST,
+};
