+++ js/main.js
@@ -1,147 +1,4 @@
-const APP_DESCRIPTION_BANK = ['Господа, сплочённость команды профессионалов однозначно фиксирует необходимость позиций, занимаемых участниками в отношении поставленных задач. Противоположная точка зрения подразумевает, что предприниматели в сети интернет, вне зависимости от их уровня, должны быть описаны максимально подробно! Имеется спорная точка зрения, гласящая примерно следующее: тщательные исследования конкурентов будут функционально разнесены на независимые элементы.', 'Господа, сложившаяся структура организации предполагает независимые способы реализации как самодостаточных, так и внешне зависимых концептуальных решений. Каждый из нас понимает очевидную вещь: новая модель организационной деятельности позволяет выполнить важные задания по разработке экспериментов, поражающих по своей масштабности и грандиозности. А ещё сделанные на базе интернет-аналитики выводы освещают чрезвычайно интересные особенности картины в целом, однако конкретные выводы, разумеется, ассоциативно распределены по отраслям!', 'Повседневная практика показывает, что постоянный количественный рост и сфера нашей активности предполагает независимые способы реализации системы массового участия. Современные технологии достигли такого уровня, что глубокий уровень погружения выявляет срочную потребность как самодостаточных, так и внешне зависимых концептуальных решений. В своём стремлении повысить качество жизни, они забывают, что выбранный нами инновационный путь, а также свежий взгляд на привычные вещи — безусловно открывает новые горизонты для экономической целесообразности принимаемых решений.', 'Как принято считать, базовые сценарии поведения пользователей неоднозначны и будут призваны к ответу. Значимость этих проблем настолько очевидна, что понимание сути ресурсосберегающих технологий однозначно фиксирует необходимость переосмысления внешнеэкономических политик. Кстати,  многие известные личности будут рассмотрены исключительно в разрезе маркетинговых и финансовых предпосылок.', 'Задача организации, в особенности же перспективное планирование требует определения и уточнения экспериментов, поражающих по своей масштабности и грандиозности. Мы вынуждены отталкиваться от того, что существующая теория требует определения и уточнения дальнейших направлений развития. Идейные соображения высшего порядка, а также постоянный количественный рост и сфера нашей активности не оставляет шанса для экономической целесообразности принимаемых решений!', 'Банальные, но неопровержимые выводы, а также активно развивающиеся страны третьего мира, инициированные исключительно синтетически, объявлены нарушающими общечеловеческие нормы этики и морали. Но реализация намеченных плановых заданий создаёт необходимость включения в производственный план целого ряда внеочередных мероприятий с учётом комплекса новых предложений. И нет сомнений, что стремящиеся вытеснить традиционное производство, нанотехнологии, превозмогая сложившуюся непростую экономическую ситуацию, рассмотрены исключительно в разрезе маркетинговых и финансовых предпосылок.', 'Предварительные выводы неутешительны: семантический разбор внешних противодействий однозначно фиксирует необходимость новых принципов формирования материально-технической и кадровой базы. Таким образом, перспективное планирование прекрасно подходит для реализации распределения внутренних резервов и ресурсов. Идейные соображения высшего порядка, а также постоянный количественный рост и сфера нашей активности требует определения и уточнения прогресса профессионального сообщества.', 'Есть над чем задуматься: реплицированные с зарубежных источников, современные исследования и по сей день остаются уделом либералов, которые жаждут быть представлены в исключительно положительном свете. Таким образом, дальнейшее развитие различных форм деятельности однозначно фиксирует необходимость вывода текущих активов. Картельные сговоры не допускают ситуации, при которой многие известные личности освещают чрезвычайно интересные особенности картины в целом, однако конкретные выводы, разумеется, объединены в целые кластеры себе подобных!', 'Высокий уровень вовлечения представителей целевой аудитории является четким доказательством простого факта: понимание сути ресурсосберегающих технологий в значительной степени обусловливает важность системы массового участия. Повседневная практика показывает, что сплочённость команды профессионалов говорит о возможностях модели развития. Предварительные выводы неутешительны: существующая теория является качественно новой ступенью приоретизации разума над эмоциями.', 'Каждый из нас понимает очевидную вещь: понимание сути ресурсосберегающих технологий способствует повышению качества стандартных подходов. Но повышение уровня гражданского сознания предполагает независимые способы реализации стандартных подходов. Наше дело не так однозначно, как может показаться: высокое качество позиционных исследований не оставляет шанса для укрепления моральных ценностей.', 'Но действия представителей оппозиции представляют собой не что иное, как квинтэссенцию победы маркетинга над разумом и должны быть в равной степени предоставлены сами себе. Являясь всего лишь частью общей картины, представители современных социальных резервов разоблачены. В своём стремлении повысить качество жизни, они забывают, что реализация намеченных плановых заданий создаёт необходимость включения в производственный план целого ряда внеочередных мероприятий с учётом комплекса приоретизации разума над эмоциями.', 'С другой стороны, современная методология разработки способствует повышению качества вывода текущих активов. Но тщательные исследования конкурентов преданы социально-демократической анафеме. Внезапно, ключевые особенности структуры проекта представляют собой не что иное, как квинтэссенцию победы маркетинга над разумом и должны быть объявлены нарушающими общечеловеческие нормы этики и морали.', 'Таким образом, перспективное планирование требует определения и уточнения вывода текущих активов. Наше дело не так однозначно, как может показаться: курс на социально-ориентированный национальный проект не даёт нам иного выбора, кроме определения поставленных обществом задач. Но предприниматели в сети интернет объективно рассмотрены соответствующими инстанциями!', 'В своём стремлении улучшить пользовательский опыт мы упускаем, что базовые сценарии поведения пользователей, превозмогая сложившуюся непростую экономическую ситуацию, разоблачены. Есть над чем задуматься: стремящиеся вытеснить традиционное производство, нанотехнологии лишь добавляют фракционных разногласий и превращены в посмешище, хотя само их существование приносит несомненную пользу обществу. Принимая во внимание показатели успешности, синтетическое тестирование в значительной степени обусловливает важность первоочередных требований!', 'Но реализация намеченных плановых заданий требует определения и уточнения поэтапного и последовательного развития общества. Мы вынуждены отталкиваться от того, что глубокий уровень погружения в значительной степени обусловливает важность экономической целесообразности принимаемых решений. Предварительные выводы неутешительны: перспективное планирование требует анализа глубокомысленных рассуждений.', 'Господа, постоянный количественный рост и сфера нашей активности прекрасно подходит для реализации как самодостаточных, так и внешне зависимых концептуальных решений. Для современного мира повышение уровня гражданского сознания предопределяет высокую востребованность вывода текущих активов. Банальные, но неопровержимые выводы, а также независимые государства призывают нас к новым свершениям, которые, в свою очередь, должны быть заблокированы в рамках своих собственных рациональных ограничений.', 'Высокий уровень вовлечения представителей целевой аудитории является четким доказательством простого факта: понимание сути ресурсосберегающих технологий напрямую зависит от системы массового участия. С учётом сложившейся международной обстановки, повышение уровня гражданского сознания не оставляет шанса для кластеризации усилий. Сложно сказать, почему тщательные исследования конкурентов, инициированные исключительно синтетически, объединены в целые кластеры себе подобных.', 'Лишь независимые государства освещают чрезвычайно интересные особенности картины в целом, однако конкретные выводы, разумеется, ассоциативно распределены по отраслям. Лишь стремящиеся вытеснить традиционное производство, нанотехнологии лишь добавляют фракционных разногласий и указаны как претенденты на роль ключевых факторов. Безусловно, новая модель организационной деятельности однозначно фиксирует необходимость дальнейших направлений развития!', 'Лишь независимые государства набирают популярность среди определенных слоев населения, а значит, должны быть объявлены нарушающими общечеловеческие нормы этики и морали. В своём стремлении улучшить пользовательский опыт мы упускаем, что действия представителей оппозиции, вне зависимости от их уровня, должны быть обнародованы. Картельные сговоры не допускают ситуации, при которой ключевые особенности структуры проекта формируют глобальную экономическую сеть и при этом —  обнародованы.', 'В своём стремлении улучшить пользовательский опыт мы упускаем, что сторонники тоталитаризма в науке призывают нас к новым свершениям, которые, в свою очередь, должны быть своевременно верифицированы. Современные технологии достигли такого уровня, что сплочённость команды профессионалов играет определяющее значение для анализа существующих паттернов поведения. И нет сомнений, что сделанные на базе интернет-аналитики выводы своевременно верифицированы.', 'Равным образом, повышение уровня гражданского сознания выявляет срочную потребность приоретизации разума над эмоциями. Внезапно, диаграммы связей могут быть обнародованы. Ясность нашей позиции очевидна: дальнейшее развитие различных форм деятельности однозначно определяет каждого участника как способного принимать собственные решения касаемо переосмысления внешнеэкономических политик.', 'В рамках спецификации современных стандартов, действия представителей оппозиции набирают популярность среди определенных слоев населения, а значит, должны быть смешаны с не уникальными данными до степени совершенной неузнаваемости, из-за чего возрастает их статус бесполезности. Как принято считать, некоторые особенности внутренней политики будут ассоциативно распределены по отраслям. Мы вынуждены отталкиваться от того, что глубокий уровень погружения обеспечивает широкому кругу (специалистов) участие в формировании как самодостаточных, так и внешне зависимых концептуальных решений.', 'Прежде всего, консультация с широким активом говорит о возможностях системы массового участия. Следует отметить, что реализация намеченных плановых заданий не даёт нам иного выбора, кроме определения поставленных обществом задач. Прежде всего, повышение уровня гражданского сознания однозначно определяет каждого участника как способного принимать собственные решения касаемо модели развития!', 'А ещё ключевые особенности структуры проекта, превозмогая сложившуюся непростую экономическую ситуацию, разоблачены. Также как укрепление и развитие внутренней структуры создаёт предпосылки для существующих финансовых и административных условий. Безусловно, понимание сути ресурсосберегающих технологий говорит о возможностях приоретизации разума над эмоциями.', 'Противоположная точка зрения подразумевает, что непосредственные участники технического прогресса, которые представляют собой яркий пример континентально-европейского типа политической культуры, будут разоблачены. С другой стороны, внедрение современных методик предоставляет широкие возможности для как самодостаточных, так и внешне зависимых концептуальных решений! Для современного мира глубокий уровень погружения не даёт нам иного выбора, кроме определения своевременного выполнения сверхзадачи.', 'Повседневная практика показывает, что сплочённость команды профессионалов требует анализа экспериментов, поражающих по своей масштабности и грандиозности. Лишь элементы политического процесса, которые представляют собой яркий пример континентально-европейского типа политической культуры, будут смешаны с не уникальными данными до степени совершенной неузнаваемости, из-за чего возрастает их статус бесполезности. Сложно сказать, почему сделанные на базе интернет-аналитики выводы неоднозначны и будут призваны к ответу.', 'Господа, внедрение современных методик обеспечивает актуальность распределения внутренних резервов и ресурсов. Как уже неоднократно упомянуто, реплицированные с зарубежных источников, современные исследования представляют собой не что иное, как квинтэссенцию победы маркетинга над разумом и должны быть указаны как претенденты на роль ключевых факторов. Предварительные выводы неутешительны: новая модель организационной деятельности влечет за собой процесс внедрения и модернизации соответствующих условий активизации.', 'Идейные соображения высшего порядка, а также консультация с широким активом представляет собой интересный эксперимент проверки кластеризации усилий! Для современного мира курс на социально-ориентированный национальный проект является качественно новой ступенью своевременного выполнения сверхзадачи. Предварительные выводы неутешительны: сплочённость команды профессионалов создаёт необходимость включения в производственный план целого ряда внеочередных мероприятий с учётом комплекса анализа существующих паттернов поведения.', 'Предварительные выводы неутешительны: реализация намеченных плановых заданий предопределяет высокую востребованность системы обучения кадров, соответствующей насущным потребностям. Предварительные выводы неутешительны: разбавленное изрядной долей эмпатии, рациональное мышление предполагает независимые способы реализации системы массового участия. Равным образом, понимание сути ресурсосберегающих технологий представляет собой интересный эксперимент проверки прогресса профессионального сообщества.', 'Внезапно, сделанные на базе интернет-аналитики выводы ассоциативно распределены по отраслям! Противоположная точка зрения подразумевает, что элементы политического процесса и по сей день остаются уделом либералов, которые жаждут быть ассоциативно распределены по отраслям. Есть над чем задуматься: ключевые особенности структуры проекта набирают популярность среди определенных слоев населения, а значит, должны быть описаны максимально подробно.', 'Для современного мира базовый вектор развития обеспечивает широкому кругу (специалистов) участие в формировании системы обучения кадров, соответствующей насущным потребностям. В рамках спецификации современных стандартов, диаграммы связей призывают нас к новым свершениям, которые, в свою очередь, должны быть превращены в посмешище, хотя само их существование приносит несомненную пользу обществу. А также интерактивные прототипы, инициированные исключительно синтетически, обнародованы.', 'Принимая во внимание показатели успешности, постоянный количественный рост и сфера нашей активности обеспечивает широкому кругу (специалистов) участие в формировании приоретизации разума над эмоциями. Вот вам яркий пример современных тенденций — консультация с широким активом влечет за собой процесс внедрения и модернизации системы обучения кадров, соответствующей насущным потребностям. В целом, конечно, реализация намеченных плановых заданий выявляет срочную потребность форм воздействия.', 'Не следует, однако, забывать, что выбранный нами инновационный путь способствует подготовке и реализации своевременного выполнения сверхзадачи! Современные технологии достигли такого уровня, что высокое качество позиционных исследований требует от нас анализа вывода текущих активов. Господа, синтетическое тестирование создаёт необходимость включения в производственный план целого ряда внеочередных мероприятий с учётом комплекса как самодостаточных, так и внешне зависимых концептуальных решений.', 'Высокий уровень вовлечения представителей целевой аудитории является четким доказательством простого факта: разбавленное изрядной долей эмпатии, рациональное мышление, а также свежий взгляд на привычные вещи — безусловно открывает новые горизонты для поэтапного и последовательного развития общества. Имеется спорная точка зрения, гласящая примерно следующее: сделанные на базе интернет-аналитики выводы функционально разнесены на независимые элементы. Значимость этих проблем настолько очевидна, что граница обучения кадров не даёт нам иного выбора, кроме определения приоретизации разума над эмоциями.', 'Господа, реализация намеченных плановых заданий является качественно новой ступенью приоретизации разума над эмоциями. Как уже неоднократно упомянуто, интерактивные прототипы и по сей день остаются уделом либералов, которые жаждут быть преданы социально-демократической анафеме. Противоположная точка зрения подразумевает, что сделанные на базе интернет-аналитики выводы и по сей день остаются уделом либералов, которые жаждут быть объявлены нарушающими общечеловеческие нормы этики и морали.', 'Принимая во внимание показатели успешности, существующая теория обеспечивает широкому кругу (специалистов) участие в формировании экономической целесообразности принимаемых решений. Следует отметить, что дальнейшее развитие различных форм деятельности позволяет оценить значение новых принципов формирования материально-технической и кадровой базы. Господа, выбранный нами инновационный путь требует от нас анализа переосмысления внешнеэкономических политик.', 'С учётом сложившейся международной обстановки, глубокий уровень погружения способствует подготовке и реализации инновационных методов управления процессами. Картельные сговоры не допускают ситуации, при которой стремящиеся вытеснить традиционное производство, нанотехнологии набирают популярность среди определенных слоев населения, а значит, должны быть указаны как претенденты на роль ключевых факторов. Высокий уровень вовлечения представителей целевой аудитории является четким доказательством простого факта: высококачественный прототип будущего проекта позволяет оценить значение экспериментов, поражающих по своей масштабности и грандиозности.', 'Значимость этих проблем настолько очевидна, что курс на социально-ориентированный национальный проект предоставляет широкие возможности для экспериментов, поражающих по своей масштабности и грандиозности. Повседневная практика показывает, что постоянный количественный рост и сфера нашей активности играет определяющее значение для экспериментов, поражающих по своей масштабности и грандиозности. Безусловно, высокотехнологичная концепция общественного уклада обеспечивает актуальность системы обучения кадров, соответствующей насущным потребностям!', 'Современные технологии достигли такого уровня, что глубокий уровень погружения однозначно фиксирует необходимость экономической целесообразности принимаемых решений. Лишь действия представителей оппозиции неоднозначны и будут смешаны с не уникальными данными до степени совершенной неузнаваемости, из-за чего возрастает их статус бесполезности. Противоположная точка зрения подразумевает, что базовые сценарии поведения пользователей и по сей день остаются уделом либералов, которые жаждут быть в равной степени предоставлены сами себе.', 'Не следует, однако, забывать, что высококачественный прототип будущего проекта является качественно новой ступенью стандартных подходов. Высокий уровень вовлечения представителей целевой аудитории является четким доказательством простого факта: реализация намеченных плановых заданий обеспечивает актуальность форм воздействия. В своём стремлении повысить качество жизни, они забывают, что экономическая повестка сегодняшнего дня выявляет срочную потребность стандартных подходов.'];
+import { useKekstagramStore } from './keks-db/index.js';
 
-const APP_NAMES_BANK = [{ firstName: 'Terry', lastName: 'Medhurst', maidenName: 'Smitham' }, { firstName: 'Sheldon', lastName: 'Quigley', maidenName: 'Cole' }, { firstName: 'Terrill', lastName: 'Hills', maidenName: 'Hoeger' }, { firstName: 'Miles', lastName: 'Cummerata', maidenName: 'Maggio' }, { firstName: 'Mavis', lastName: 'Schultz', maidenName: 'Yundt' }, { firstName: 'Alison', lastName: 'Reichert', maidenName: 'Franecki' }, { firstName: 'Oleta', lastName: 'Abbott', maidenName: 'Wyman' }, { firstName: 'Ewell', lastName: 'Mueller', maidenName: 'Durgan' }, { firstName: 'Demetrius', lastName: 'Corkery', maidenName: 'Gleason' }, { firstName: 'Eleanora', lastName: 'Price', maidenName: 'Cummings' }, { firstName: 'Marcel', lastName: 'Jones', maidenName: 'Smith' }, { firstName: 'Assunta', lastName: 'Rath', maidenName: 'Heller' }, { firstName: 'Trace', lastName: 'Douglas', maidenName: 'Lemke' }, { firstName: 'Enoch', lastName: 'Lynch', maidenName: 'Heidenreich' }, { firstName: 'Jeanne', lastName: 'Halvorson', maidenName: 'Cummerata' }, { firstName: 'Trycia', lastName: 'Fadel', maidenName: 'Rosenbaum' }, { firstName: 'Bradford', lastName: 'Prohaska', maidenName: 'Bins' }, { firstName: 'Arely', lastName: 'Skiles', maidenName: 'Monahan' }, { firstName: 'Gust', lastName: 'Purdy', maidenName: 'Abshire' }, { firstName: 'Lenna', lastName: 'Renner', maidenName: 'Schumm' }, { firstName: 'Doyle', lastName: 'Ernser', maidenName: 'Feeney' }, { firstName: 'Tressa', lastName: 'Weber', maidenName: 'Williamson' }, { firstName: 'Felicity', lastName: 'O`Reilly', maidenName: 'Rosenbaum' }, { firstName: 'Jocelyn', lastName: 'Schuster', maidenName: 'Dooley' }, { firstName: 'Edwina', lastName: 'Ernser', maidenName: 'Kiehn' }, { firstName: 'Griffin', lastName: 'Braun', maidenName: 'Deckow' }, { firstName: 'Piper', lastName: 'Schowalter', maidenName: 'Wuckert' }, { firstName: 'Kody', lastName: 'Terry', maidenName: 'Larkin' }, { firstName: 'Macy', lastName: 'Greenfelder', maidenName: 'Koepp' }, { firstName: 'Maurine', lastName: 'Stracke', maidenName: 'Abshire' }];
-
-const APP_COMMENTS_BANK = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const APP_PHOTO_COUNT = 25;
-const APP_USERS_COUNT = 6;
-const APP_COMMENTS_FOR_POST = 5;
-
-/**
- * @typedef Comment
- * @property {number} id
- * @property {string} avatar
- * @property {string} message
- * @property {string} name
- * @description Комментарий пользователя
- *
- * @typedef Photo
- * @property {number} id
- * @property {string } url
- * @property {string} description
- * @property {number} likes
- * @property {Array<Comment>} comments
- * @description Фото пользователя
- */
-
-const getRandomInt = (min = 0, max = Number.MAX_SAFE_INTEGER) =>
-  Math.abs(Math.floor(Math.random() * (max - min + 1) + min));
-
-const createSequnce = () => {
-  const seqReg = [];
-  let seqHead = 0;
-
-  const nextValue = () => {
-    seqHead++;
-    seqReg.push(seqHead);
-    return seqHead;
-  };
-  const currentValue = () => seqHead;
-  const exists = (val) => seqReg.includes(val);
-  return {
-    nextValue,
-    currentValue,
-    exists
-  };
-};
-const createObjectStore = (factory, count) => {
-  const idSeq = createSequnce();
-  const storeFactory = () => Object.fromEntries(
-    Array.from({ length: count }, () => {
-      const id = idSeq.nextValue();
-      return [id, factory(id)];
-    })
-  );
-  const store = storeFactory();
-
-  const selected = [];
-  const addSelected = (obj) => selected.push(obj);
-  const isAllSelected = () => selected.length >= idSeq.currentValue();
-  const cleanupSelected = () => selected.splice(0, selected.length);
-  const selectOne = () => {
-    let result = null;
-    if (!isAllSelected()) {
-      let randKey = getRandomInt(1, idSeq.currentValue());
-      while (selected.includes(randKey)) {
-        randKey = getRandomInt(1, idSeq.currentValue());
-      }
-      result = store[randKey];
-      addSelected(result);
-    }
-    return result;
-  };
-
-  const getRandom = () => store[getRandomInt(1, idSeq.currentValue())];
-  const getRandomBatch = (batchCount) => {
-    const batch = [];
-    while (batch.length < batchCount && batch.length < idSeq.currentValue()) {
-      batch.push(getRandom());
-    }
-    return batch;
-  };
-  const selectQuery = () => Object.values(store);
-  return {
-    selectQuery,
-    selectOne,
-    cleanupSelected,
-    getRandom,
-    getRandomBatch
-  };
-};
-
-const usersFactory = () => {
-  const namesStore = createObjectStore((id) => APP_NAMES_BANK[id - 1], APP_NAMES_BANK.length + 1, true);
-  return createObjectStore(
-    (id) => ({
-      id,
-      ...namesStore.getRandom(),
-      avatar: `img/avatar-${id}.svg`
-    }),
-    APP_USERS_COUNT
-  );
-};
-const commentsFactory = (users) => {
-  const commentStore = createObjectStore((id) => APP_COMMENTS_BANK[id - 1], APP_COMMENTS_BANK.length + 1, true);
-  return createObjectStore(
-    (id) => {
-      const user = users.getRandom();
-      return {
-        id,
-        avatar: user.avatar,
-        name: user.firstName,
-        message: commentStore.getRandom(),
-      };
-    },
-    APP_PHOTO_COUNT * APP_COMMENTS_FOR_POST
-  );
-};
-
-const photosFactory = (comments) => {
-  const descriptions = createObjectStore((id) => APP_DESCRIPTION_BANK[id - 1], APP_DESCRIPTION_BANK.length + 1, true);
-  return createObjectStore(
-    (id) => ({
-      id,
-      url: `photos/${id}.jpg`,
-      description: descriptions.getRandom(),
-      likes: getRandomInt(0,1000),
-      comments: comments.getRandomBatch(getRandomInt(1, APP_COMMENTS_FOR_POST))
-    }),
-    APP_PHOTO_COUNT
-  );
-};
-
-// eslint-disable-next-line no-unused-vars
-const generateKekstagram = () => {
-  const users = usersFactory();
-  const comments = commentsFactory(users);
-  const photo = photosFactory(comments);
-  return photo.selectQuery();
-};
+const kekstagramStore = useKekstagramStore();
+console.log(kekstagramStore.getPhotos());
